+++ reqbaz/src/main/java/de/rwth/dbis/acis/bazaar/service/dal/entities/Attachment.java
@@ -73,4 +73,9 @@ public class Attachment extends EntityBase implements Ownable {
     public boolean isOwner(User user) {
         return creator == user;
     }
+
+    @Override
+    public boolean isOwner(Integer userId) {
+        return creator.getId() == userId;
+    }
 }