+++ reqbaz/src/main/java/de/rwth/dbis/acis/bazaar/service/dal/entities/Category.java
@@ -75,4 +75,9 @@ public class Category extends EntityBase implements Ownable {
     public boolean isOwner(User user) {
         return creator == user;
     }
+
+    @Override
+    public boolean isOwner(Integer userId) {
+        return creator.getId() == userId;
+    }
 }