+++ reqbaz/src/main/java/de/rwth/dbis/acis/bazaar/service/dal/entities/Requirement.java
@@ -76,4 +76,9 @@ public class Requirement extends EntityBase implements Ownable {
     public boolean isOwner(User user) {
         return creator == user;
     }
+
+    @Override
+    public boolean isOwner(Integer userId) {
+        return creator.getId() == userId;
+    }
 }