+++ reqbaz/src/test/java/de/rwth/dbis/acis/bazaar/service/dal/DALFacadeTest.java
@@ -159,6 +159,62 @@ public class DALFacadeTest extends SetupData {
 
     }
 
+    @Test
+    public void testCreateCategory() throws Exception {
+        String name = "TestComp9";
+        Category testComp9 = Category.builder()
+                .name(name)
+                .description("Very testing")
+                .projectId(testProject.getId())
+                .creator(initUser)
+                .build();
+
+        facade.createCategory(testComp9, initUser.getId());
+
+        List<Category> components = facade.listCategoriesByProjectId(testProject.getId(), ALL_IN_ONE_PAGE, initUser.getId()).getElements();
+
+        assertNotNull(components);
+        assertEquals(2,components.size());
+
+        components.forEach(category -> {
+            if (category.getName().equals(name)) {
+                jooq.delete(de.rwth.dbis.acis.bazaar.dal.jooq.tables.Category.CATEGORY).where(de.rwth.dbis.acis.bazaar.dal.jooq.tables.Category.CATEGORY.ID.equal(category.getId())).execute();
+            }
+        });
+    }
+
+    @Test
+    public void testComments() throws Exception {
+        Comment testComment = Comment.builder()
+                .message("TestComment")
+                .requirementId(testRequirement.getId())
+                .creator(initUser)
+                .build();
+
+        Comment createdComment = facade.createComment(testComment);
+
+        List<Comment> comments = facade.listCommentsByRequirementId(testRequirement.getId(), ALL_IN_ONE_PAGE).getElements();
+
+        assertNotNull(comments);
+        assertEquals(1,comments.size());
+        assertEquals(createdComment.getId(), comments.get(0).getId());
+
+        createdComment.setMessage("Updated message");
+
+        Comment updatedComment = facade.updateComment(createdComment);
+
+        assertNotNull(updatedComment);
+        assertEquals("Updated message", updatedComment.getMessage());
+
+        facade.deleteCommentById(updatedComment.getId());
+
+        comments = facade.listCommentsByRequirementId(testRequirement.getId(), ALL_IN_ONE_PAGE).getElements();
+        assertNotNull(comments);
+        assertEquals(0,comments.size());
+
+
+    }
+
     /* Doesn't work, searching breaks
     public void testModifyProject() throws Exception {
         //TODO
@@ -287,63 +343,12 @@ public class DALFacadeTest extends SetupData {
         assertEquals(1,components.get(0).getId());
     }
 
-    public void testCreateCategory() throws Exception {
-        int createdComponentId = 9;
-        Category testComp9 = Category.getBuilder("TestComp9").description("Very testing").id(createdComponentId).projectId(1).leader(initUser).build();
-
-        facade.createCategory(testComp9, initUser.getId());
-
-        List<Category> components = (List<Category>) facade.listCategoriesByProjectId(1, ALL_IN_ONE_PAGE, 1);
-
-        assertNotNull(components);
-        assertEquals(1,components.size());
-        assertEquals(createdComponentId,components.get(0).getId());
-
-        jooq.delete(de.rwth.dbis.acis.bazaar.dal.jooq.tables.Category.CATEGORY).where(de.rwth.dbis.acis.bazaar.dal.jooq.tables.Category.CATEGORY.ID.equal(createdComponentId)).execute();
-    }
-
     public void testModifyComponent() throws Exception {
         //TODO
     }
 
     public void testDeleteComponentById() throws Exception {
         //TODO
-    }
-
-    public void testListCommentsByRequirementId() throws Exception {
-        List<Comment> comments = facade.listCommentsByRequirementId(2, ALL_IN_ONE_PAGE).getElements();
-
-        assertNotNull(comments);
-        assertEquals(2,comments.size());
-        assertTrue(comments.get(0).getId() == 1 || comments.get(0).getId() == 2);
-        assertTrue(comments.get(1).getId() == 1 || comments.get(1).getId() == 2);
-    }
-
-    public void testCreateComment() throws Exception {
-        int createdCommentId = 9;
-        Comment testComment = Comment.getBuilder("TestComment").id(createdCommentId).creator(initUser).requirementId(1).build();
-
-        facade.createComment(testComment);
-
-        List<Comment> comments = facade.listCommentsByRequirementId(1, ALL_IN_ONE_PAGE).getElements();
-
-        assertNotNull(comments);
-        assertEquals(1,comments.size());
-        assertEquals(createdCommentId,comments.get(0).getId());
-
-        jooq.delete(de.rwth.dbis.acis.bazaar.dal.jooq.tables.Comment.COMMENT).where(de.rwth.dbis.acis.bazaar.dal.jooq.tables.Comment.COMMENT.ID.equal(createdCommentId)).execute();
-    }
-
-    public void testDeleteCommentById() throws Exception {
-        Comment testComment = Comment.getBuilder("TestComment").id(9).creator(initUser).requirementId(1).build();
-
-        facade.createComment(testComment);
-
-        facade.deleteCommentById(9);
-
-        List<Comment> comments = facade.listCommentsByRequirementId(1, ALL_IN_ONE_PAGE).getElements();
-        assertNotNull(comments);
-        assertEquals(0, comments.size());
     }
 
     public void testVoteAndUnvote() throws Exception {